stage4: build packages with C.utf8 Ctype
Some python packages require a utf8 ctype when opening files during
compilation. While these packages are not required for stage123,
there are a handful which might be pulled in as additional packages
in stage4/embedded.

Until a more permanent fix is in place, set LC_CTYPE to a utf8 value.

https://bugs.gentoo.org/688330
https://devmanual.gentoo.org/eclass-reference/python-utils-r1.eclass/index.html#lbAE