Loosen commandline argument parser
For arguments given on the command line, use ParserBase instead of SpecParser.

With SpecParser, an empty directive is considered unnecessary and is
removed from the settings object.

This hampers us if we need to run --depclean to reduce the size of our
stages, but don't have any specific packages to remove once the stage is
clean.

Depcleans are run as part of the `controller.sh unmerge` script, but
this script is only run if "<target>/unmerge" is in the settings object.