Use a 3 stage livecd build
The current (2 stage) livecd process requires a full stage3/4 tarball
that can install genkernel and compile a kernel for the livecd. Once the
kernel is built, packages are then unmerged to reduce the size of the CD
image. This means explicity removing every unnecessary package that may
be installed from the system set, or additional stage4 world sets.

Instead, an embedded target could be used, or a stage1 target be
abused, to create minimal images for the livecd environment.

This patch moves some actions from livecd-stage1 and livecd-stage2 into
a new target: livecd-stage3.

livecd-stage1 is now very similar to a regular stage4, where additional
packages can be installed through the package manager. This step would
be optional for embedded stages without a package manager.

livecd-stage2 now uses a stage3 to build a kernel and reusable "cdtar".
The seed stage is not retained.

livecd-stage3 takes the built kernel and modules, and a
livecd-stage1/embedded/stage3/stage4 tarball. Without package manager
interaction, it combines these into a livecd. Modules for the built
kernels are extracted into the looped image.

This patch is an RFC as it will break a lot of existing specfiles, and a
fsscript would also be necessary for the resulting livecd to be useful.

The code basically copies the existing livecd-stage1/livecd-stage2 code
in the livecd-stage3 target and support scripts. I'm sure this could be
cleaned up to remove the references in livecd-stage1.py/livecd-stage2.py

The livecd-stage3 specfile requires the 'profile:' keyword, even though
it is not required. This is because it is defined as a requirement by
StageBase.__init__()

livecd_stage3: allow fsops param

livecd-stage3: reorder steps

cleanup machine-id file

livecd: remove unbind step

livecd-stage3: convert action sequence to new-style

livecd-stage3: fix bind in its own mount namespace

livecd: fix bootargs in the cdtar

Fix livecd-stage3 action sequence method references

livecd-stage3: remove existing build_sequence steps

livecd-stage3: skip sequences that require a package manager

patch livecd-stage2-cleanup-stages

patch livecd-stage1-run-unmerge-step

patch step3-profile-required